requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
Django==2.1.3
django-extensions==2.1.3
djangorestframework==3.9.0
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
numpy==1.15.4
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psycopg2==2.7.3
ptyprocess==0.6.0
pudb==2018.1
Pygments==2.2.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
spotipy==2.4.4
traitlets==4.3.2
urllib3==1.24.1
urwid==2.0.1
wcwidth==0.1.7
